src/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic
torch 
torchvision 
torchaudio --index-url https://download.pytorch.org/whl/cpu
transformers
sentencepiece
sacremoses